Fix rebase issues + lint issues
