+++ client/src/panels/panel_declarations/finances/auth_exp_planned_spending/auth_exp_planned_spending.js
@@ -1,5 +1,6 @@
 import { Fragment } from "react";
 
+import { RotateLandscape } from "../../../../components";
 import {
   run_template,
   year_templates,
@@ -14,7 +15,6 @@ import {
   util_components,
   declare_panel,
 } from "../../shared.js";
-import { RotateLandscape } from "../../../../components";
 
 import text from "./auth_exp_planned_spending.yaml";
 import "./auth_exp_planned_spending.scss";
