+++ client/src/panels/panel_declarations/finances/sobj/personel_spend.js
@@ -1,3 +1,4 @@
+import { RotateLandscape } from "../../../../components";
 import {
   declare_panel,
   year_templates,
@@ -11,7 +12,6 @@ import {
 } from "../../shared.js";
 
 import { text_maker, TM } from "./sobj_text_provider.js";
-import { RotateLandscape } from "../../../../components";
 const { sos } = businessConstants;
 const { std_years } = year_templates;
 const { Format } = util_components;
